added results page

+++ UserSide/public/js/logout.js
@@ -1,4 +1,10 @@
 $(document).on("click", "#logout", (e) => {
-    window.location.href = "http://localhost:4000/";
-    localStorage.clear();
+    const token = localStorage.getItem('access-token');
+    if (!token) {
+        window.location.href = "http://localhost:4000/"
+    } else {
+        window.location.href = "http://localhost:4000/";
+        localStorage.clear();
+    }
+
 })
